backend/requirements.txt
fastapi==0.124.4
uvicorn==0.38.0
beautifulsoup4==4.14.3
requests==2.32.5
networkx==3.6.1
pydantic==2.12.5
python-multipart==0.0.21
httpx==0.28.1
numpy==2.3.5
pandas==2.3.3
openpyxl==3.1.5
python-jose[cryptography]==3.5.0
passlib==1.7.4
bcrypt==3.2.0
python-dotenv==1.2.1
motor==3.7.1
pymongo==4.15.5
duckduckgo-search==7.1.0
selenium==4.27.1
webdriver-manager==4.0.2